Harden cross-platform release tooling
