+++ .gitattributes
@@ -1,0 +1,1 @@
+release/release-lock.json text eol=lf
